Fix prospection data task
