+++ src/main/java/com/gregtechceu/gtceu/common/commands/GTClientCommands.java
@@ -5,8 +5,6 @@ import net.minecraft.commands.CommandSourceStack;
 
 import com.mojang.brigadier.CommandDispatcher;
 
-import static net.minecraft.commands.Commands.*;
-
 public class GTClientCommands {
 
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher, CommandBuildContext buildContext) {}
